pyproject.toml
[tool.poetry]
name = "trollbot"
version = "0.1.0"
description = "A wrapper library dedicated to making bots for Trollbox (and chatrooms alike)"
authors = ["mewhenthe <xsare435@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-socketio = "4.6.1"

[tool.poetry.scripts]
sample = "examples.01_sample:sample"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
